+++ src/dataregistry/query.py
@@ -106,7 +106,14 @@ class Query:
         self._metadata = TableMetadata(db_connection)
 
         # Get table definitions
-        self._table_list = ["dataset", "execution", "dataset_alias", "dependency"]
+        self._table_list = [
+            "dataset",
+            "execution",
+            "dataset_alias",
+            "dependency",
+            "keyword",
+            "dataset_keyword",
+        ]
         self._get_database_tables()
 
     def get_all_columns(self):
@@ -317,7 +324,9 @@ class Query:
 
         # No properties requested, return all from dataset table (only)
         if property_names is None:
-            stmt = select("*").select_from(self._tables["dataset"])
+            #stmt = select("*").select_from(self._tables["dataset"])
+            #stmt = select("*").select_from(self._tables["keyword"])
+            stmt = select(self._tables["dataset"]).join(self._tables["dataset_keyword"])
 
         # Return the selected properties.
         else: